Bone structure for the phpsec\framework
Created the framework folder, under it the _core folder which will have
the framework code, the config, control, model, static, test and view
folders which will contain the application files.

under _core, the loader.php handles environment setup, autoload.php
setups the autoloader for the framework and the phpsec libs, and
front.php is the front-controller which only shows envars right now.